chore: refactor assets
