+++ crates/bevy_animation_graph/src/core/animation_clip/mod.rs
@@ -1,3 +1,5 @@
+pub mod loader;
+
 use bevy::{
     animation::{AnimationCurves, AnimationTargetId},
     asset::{prelude::*, ReflectAsset},
