+++ crates/bevy_animation_graph/src/core/plugin.rs
@@ -1,14 +1,16 @@
+use super::animation_clip::loader::GraphClipLoader;
 use super::animation_clip::Interpolation;
 use super::edge_data::{AnimationEvent, EventQueue, SampledEvent};
 use super::pose::Pose;
+use super::prelude::loader::AnimatedSceneLoader;
 use super::prelude::{locate_animated_scene_player, GraphClip};
 use super::skeleton::loader::SkeletonLoader;
 use super::skeleton::Skeleton;
 use super::state_machine::high_level::GlobalTransition;
 use super::systems::apply_animation_to_targets;
 use super::{
-    animated_scene::{spawn_animated_scenes, AnimatedScene, AnimatedSceneLoader},
-    animation_graph::loader::{AnimationGraphLoader, GraphClipLoader},
+    animated_scene::{spawn_animated_scenes, AnimatedScene},
+    animation_graph::loader::AnimationGraphLoader,
     edge_data::{BoneMask, DataSpec, DataValue},
     state_machine::high_level::{loader::StateMachineLoader, StateMachine},
     systems::{animation_player, animation_player_deferred_gizmos},
